size_t is the result from the sizeof operation and is not the largest value type, the largest value type is uintmax_t.

+++ include/rock/array.h
@@ -19,8 +19,8 @@
 struct rock_array {
     void *data;
     size_t size;
-    size_t length;
-    size_t capacity;
+    uintmax_t length;
+    uintmax_t capacity;
 };
 
 /**
@@ -34,13 +34,15 @@ struct rock_array {
  * @throws ROCK_ARRAY_ERROR_MEMORY_ALLOCATION_FAILED if there is
  * insufficient memory to initialize the array instance.
  */
-bool rock_array_init(struct rock_array *object, size_t size, size_t capacity);
+bool
+rock_array_init(struct rock_array *object, size_t size, uintmax_t capacity);
 
 /**
  * @brief Invalidate the array.
  * <p>All the items contained within the array will have the given <i>on
  * destroy</i> callback invoked upon it. The actual <u>array instance
- * is not deallocated</u> since it may have been embedded in a larger structure.
+ * is not deallocated</u> since it may have been embedded in a larger
+ * structure.</p>
  * @param [in] object instance to be invalidated.
  * @param [in] on_destroy called just before the item is to be destroyed.
  * @return On success true, otherwise false if an error has occurred.
@@ -57,7 +59,7 @@ bool rock_array_invalidate(struct rock_array *object,
  * @throws ROCK_ARRAY_ERROR_OBJECT_IS_NULL if object is <i>NULL</i>.
  * @throws ROCK_ARRAY_ERROR_OUT_IS_NULL if out is <i>NULL</i>.
  */
-bool rock_array_capacity(struct rock_array *object, size_t *out);
+bool rock_array_capacity(struct rock_array *object, uintmax_t *out);
 
 /**
  * @brief Retrieve the length.
@@ -67,7 +69,7 @@ bool rock_array_capacity(struct rock_array *object, size_t *out);
  * @throws ROCK_ARRAY_ERROR_OBJECT_IS_NULL if object is <i>NULL</i>.
  * @throws ROCK_ARRAY_ERROR_OUT_IS_NULL if out is <i>NULL</i>.
  */
-bool rock_array_get_length(struct rock_array *object, size_t *out);
+bool rock_array_get_length(struct rock_array *object, uintmax_t *out);
 
 /**
  * @brief Set the length.
@@ -78,7 +80,7 @@ bool rock_array_get_length(struct rock_array *object, size_t *out);
  * @throws ROCK_ARRAY_ERROR_MEMORY_ALLOCATION_FAILED if there is not enough
  * memory to set the array length.
  */
-bool rock_array_set_length(struct rock_array *object, size_t length);
+bool rock_array_set_length(struct rock_array *object, uintmax_t length);
 
 /**
  * @brief Retrieve the size of an item.
@@ -121,7 +123,8 @@ bool rock_array_add(struct rock_array *object, void *item);
  * @throws ROCK_ARRAY_ERROR_MEMORY_ALLOCATION_FAILED if there is not enough
  * memory to all the items.
  */
-bool rock_array_add_all(struct rock_array *object, size_t count, void *items[]);
+bool rock_array_add_all(struct rock_array *object, uintmax_t count,
+                        void *items[]);
 
 /**
  * @brief Remove last item.
@@ -144,7 +147,7 @@ bool rock_array_remove_last(struct rock_array *object);
  * @throws ROCK_ARRAY_ERROR_MEMORY_ALLOCATION_FAILED if there is not enough
  * memory to add another item.
  */
-bool rock_array_insert(struct rock_array *object, size_t at, void *item);
+bool rock_array_insert(struct rock_array *object, uintmax_t at, void *item);
 
 /**
  * @brief Insert all the items at index.
@@ -157,12 +160,12 @@ bool rock_array_insert(struct rock_array *object, size_t at, void *item);
  * @throws ROCK_ARRAY_ERROR_INDEX_IS_OUT_OF_BOUNDS if at does not refer to an
  * item contained within the array.
  * @throws ROCK_ARRAY_ERROR_COUNT_IS_ZERO if count is zero.
- * @throws ROCK_ARRAY_ERROR_ITEMS_IS_NULL is items is <i>NULL</i>.
+ * @throws ROCK_ARRAY_ERROR_ITEMS_IS_NULL if items is <i>NULL</i>.
  * @throws ROCK_ARRAY_ERROR_MEMORY_ALLOCATION_FAILED if there is not enough
  * memory to add all the items to the array.
  */
-bool rock_array_insert_all(struct rock_array *object, size_t at, size_t count,
-        void *items[]);
+bool rock_array_insert_all(struct rock_array *object, uintmax_t at,
+                           uintmax_t count, void *items[]);
 
 /**
  * @brief Remove an item at the given index.
@@ -173,7 +176,7 @@ bool rock_array_insert_all(struct rock_array *object, size_t at, size_t count,
  * @throws ROCK_ARRAY_ERROR_INDEX_IS_OUT_OF_BOUNDS if at does not refer to an
  * item contained within the array.
  */
-bool rock_array_remove(struct rock_array *object, size_t at);
+bool rock_array_remove(struct rock_array *object, uintmax_t at);
 
 /**
  * @brief Remove all the items from the given index up to count items.
@@ -186,7 +189,8 @@ bool rock_array_remove(struct rock_array *object, size_t at);
  * @throws ROCK_ARRAY_ERROR_INDEX_IS_OUT_OF_BOUNDS if at does not refer to an
  * item contained within the array.
  */
-bool rock_array_remove_all(struct rock_array *object, size_t at, size_t count);
+bool rock_array_remove_all(struct rock_array *object, uintmax_t at,
+                           uintmax_t count);
 
 /**
  * @brief Retrieve the item at the given index.
@@ -199,7 +203,7 @@ bool rock_array_remove_all(struct rock_array *object, size_t at, size_t count);
  * @throws ROCK_ARRAY_ERROR_INDEX_IS_OUT_OF_BOUNDS if at does not refer to an
  * item contained within the array.
  */
-bool rock_array_get(struct rock_array *object, size_t at, void **out);
+bool rock_array_get(struct rock_array *object, uintmax_t at, void **out);
 
 /**
  * @brief Set the item at the given index.
@@ -209,11 +213,10 @@ bool rock_array_get(struct rock_array *object, size_t at, void **out);
  * array unless it is <i>NULL</i> then that index's contents will be zeroed out.
  * @return On success true, otherwise false if an error has occurred.
  * @throws ROCK_ARRAY_ERROR_OBJECT_IS_NULL if object is <i>NULL</i>.
- * @throws ROCK_ARRAY_ERROR_OUT_IS_NULL if out is <i>NULL</i>.
  * @throws ROCK_ARRAY_ERROR_INDEX_IS_OUT_OF_BOUNDS if at does not refer to an
  * item contained within the array.
  */
-bool rock_array_set(struct rock_array *object, size_t at, const void *item);
+bool rock_array_set(struct rock_array *object, uintmax_t at, const void *item);
 
 /**
  * @brief First item of the array.
